Compress cli tests into one file
